requirements.txt
langchain==0.3.25
langchain_google_genai==2.1.5
protobuf==6.31.1
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.1
unstructured==0.17.2
